+++ lib/src/state.dart
@@ -253,8 +253,16 @@ class BishopState {
 class StateMeta {
   final BuiltVariant variant;
   final MoveMeta? moveMeta;
-  const StateMeta({required this.variant, this.moveMeta});
+  final List<Iterable<int>>? checks;
+
+  const StateMeta({
+    required this.variant,
+    this.moveMeta,
+    this.checks,
+  });
 
   String? get algebraic => moveMeta?.algebraic;
   String? get prettyName => moveMeta?.formatted;
+  List<bool?> get inCheck =>
+      checks?.map((e) => e.isNotEmpty).toList() ?? [null, null];
 }