+++ test/perft_benchmark.dart
@@ -79,6 +79,7 @@ void main(List<String> args) {
       'mean NPS: ${npsAvg.toStringAsFixed(2)}');
   print('Best: $npsMax');
   print('Worst: $npsMin');
+  // todo: figure out wy the first couple of runs of each perft is so slow
 }
 
 // 12/01/23, on windows machine
@@ -89,3 +90,15 @@ void main(List<String> args) {
 // 2271406.50us [rnbq1k1r/pp1Pbppp/2p5/8/2B5/8/PPP1NnPP/RNBQK2R w KQ - 1 8, standard]
 // 3559100.50us [r4rk1/1pp1qppp/p1np1n2/2b1p1B1/2B1P1b1/P1NP1N2/1PP1QPPP/R4RK1 w - - 0 10, standard]
 // -- Total Runtime: 7747.34ms --
+
+// 28/03/23, on same machine
+// 164065.67us [rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1, chess]
+// 36043.14us [r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1, chess]
+// 771039.67us [8/2p5/3p4/KP5r/1R3p1k/8/4P1P1/8 w - - 0 1, chess]
+// 184071.36us [r3k2r/Pppp1ppp/1b3nbN/nP6/BBP1P3/q4N2/Pp1P2PP/R2Q1RK1 w kq - 0 1, chess]
+// 1078279.00us [rnbq1k1r/pp1Pbppp/2p5/8/2B5/8/PPP1NnPP/RNBQK2R w KQ - 1 8, chess]
+// 1662813.50us [r4rk1/1pp1qppp/p1np1n2/2b1p1B1/2B1P1b1/P1NP1N2/1PP1QPPP/R4RK1 w - - 0 10, chess]
+// -- Total Runtime: 3896.31ms --
+// Total Tests: 1197, mean NPS: 557866.12
+// Best: 1019500.00nps (2039/0.00s) [r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1]
+// Worst: 66432.84nps (8902/0.13s) [rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1]