+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Various minor improvements to PGN import/export.
 - Improvements to `example.dart`: can export PGN and use the engine.
 - Fixed some variants not having the 100 half-move rule when they should.
+- Variants: Clobber (`MiscVariants.clobber()`) and Clobber 10 (`MiscVariants.clobber10()`).
 
 ### 1.3.1
 - Further optimisation: excluding piece types early from attack checks (~20% performance boost).