+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.3.3
+- King attacks are calculated in advance for mainline moves and can be accessed with `state.meta!.checks!`. They come in the form of a list of squares containing pieces that are attacking the king of each player.
+
 ### 1.3.2
 - Fixed a castling bug (queenside allowed with piece on b1).
 - Various minor improvements to PGN import/export.